Feat/virgo/dev

+++ src/main/java/com/virgo/rekomendasos/utils/dto/LogTransactionDto.java
@@ -11,7 +11,7 @@ public class LogTransactionDto {
 
     private String orderId;
 
-    private Integer grossAmount;
+    private Long grossAmount;
 
     private String status;
